ref: Get rid of `utils` folder and move into `lib`
