+++ app/(app)/dashboard/components/StickyNoteDialog.tsx
@@ -3,7 +3,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from '@/components/ui/dialog';
-import type { Note, Task } from '@/utils/supabase/rows';
+import type { Note, Task } from '@/lib/supabase/rows';
 
 // Dialog component for rendering the full content of a sticky note on the dashboard when clicked
 type Props = {
