+++ app/(app)/flames/components/CreateFlameDialog.tsx
@@ -29,8 +29,8 @@ import {
   type CreateFlameFormData,
   createFlameSchema,
 } from '@/lib/schemas/flame';
+import type { Flame } from '@/lib/supabase/rows';
 import { cn } from '@/lib/utils';
-import type { Flame } from '@/utils/supabase/rows';
 import { createFlame, updateFlame } from '../actions';
 import {
   FLAME_GRADIENT_CLASSES,
