+++ app/(app)/flames/components/FlamesList.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { useEffect, useRef, useState } from 'react';
+import type { Flame, FlameSession } from '@/lib/supabase/rows';
 import { cn } from '@/lib/utils';
-import type { Flame, FlameSession } from '@/utils/supabase/rows';
 import type { FuelBudgetStatus } from '../actions';
 import { FlamesProvider, useFlamesContext } from '../hooks/useFlames';
 import { FlamesPageActions } from './FlamesPageActions';
