+++ app/(app)/flames/schedule/components/DayRow.tsx
@@ -15,8 +15,8 @@ import { useCallback, useMemo, useState } from 'react';
 import { toast } from 'sonner';
 import { FlameCard } from '@/app/(app)/flames/components/flame-card/FlameCard';
 import { Button } from '@/components/ui/button';
+import type { Flame } from '@/lib/supabase/rows';
 import { cn, parseLocalDate } from '@/lib/utils';
-import type { Flame } from '@/utils/supabase/rows';
 import { setDaySchedule } from '../actions';
 import type { DayPlan } from '../queries';
 import { formatMinutes } from '../utils';
