+++ app/(app)/flames/schedule/components/dialog/FuelSlider.tsx
@@ -7,8 +7,8 @@ import {
   FLAME_HEX_COLORS,
   getFlameColors,
 } from '@/app/(app)/flames/utils/colors';
+import type { Flame } from '@/lib/supabase/rows';
 import { cn } from '@/lib/utils';
-import type { Flame } from '@/utils/supabase/rows';
 
 const MAX_MINUTES = 720; // 12 hours
 const SNAP_INCREMENT = 15;
