+++ proxy.ts
@@ -1,5 +1,5 @@
 import type { NextRequest } from 'next/server';
-import { updateSession } from '@/utils/supabase/proxy';
+import { updateSession } from '@/lib/supabase/proxy';
 
 export async function proxy(request: NextRequest) {
   return await updateSession(request);
